[pkg/config/remote] Move api, meta and uptane to `internal`
