+++ pkg/config/remote/internal/meta/1.config.json
