+++ pkg/config/remote/internal/meta/1.director.json
