+++ pkg/config/remote/internal/uptane/local_store.go
@@ -11,7 +11,7 @@ import (
 	"encoding/json"
 	fmt "fmt"
 
-	"github.com/DataDog/datadog-agent/pkg/config/remote/meta"
+	"github.com/DataDog/datadog-agent/pkg/config/remote/internal/meta"
 	"go.etcd.io/bbolt"
 )
