+++ pkg/config/remote/internal/uptane/local_store_test.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/config/remote/meta"
+	"github.com/DataDog/datadog-agent/pkg/config/remote/internal/meta"
 	"github.com/stretchr/testify/assert"
 	"go.etcd.io/bbolt"
 )
